Separate out EventHub into its own module
