+++ azure-functions-sdk/src/commands/new.rs
@@ -18,12 +18,14 @@ use syn::{parse::Parser, parse_file, punctuated::Punctuated, Ident, Item, Token}
 
 mod blob;
 mod event_grid;
+mod event_hub;
 mod http;
 mod queue;
 mod timer;
 
 pub use self::blob::Blob;
 pub use self::event_grid::EventGrid;
+pub use self::event_hub::EventHub;
 pub use self::http::Http;
 pub use self::queue::Queue;
 pub use self::timer::Timer;
@@ -215,62 +217,6 @@ impl<'a> From<&'a ArgMatches<'a>> for New<'a> {
             quiet: args.is_present("quiet"),
             color: args.value_of("color"),
             args,
-        }
-    }
-}
-
-struct EventHub<'a> {
-    name: &'a str,
-    connection: &'a str,
-    hub_name: &'a str,
-}
-
-impl<'a> EventHub<'a> {
-    pub fn create_subcommand<'b>() -> App<'a, 'b> {
-        SubCommand::with_name("event-hub")
-            .about("Creates a new Event Hub triggered Azure Function.")
-            .arg(
-                Arg::with_name("name")
-                    .long("name")
-                    .short("n")
-                    .value_name("NAME")
-                    .help("The name of the new Azure Function.")
-                    .required(true),
-            )
-            .arg(
-                Arg::with_name("connection")
-                    .long("connection")
-                    .short("c")
-                    .value_name("CONNECTION")
-                    .help("The name of the connection setting to use for the Event Hub trigger.")
-                    .required(true),
-            )
-            .arg(
-                Arg::with_name("hub_name")
-                    .long("hub-name")
-                    .short("h")
-                    .value_name("HUBNAME")
-                    .help("The name of the Event Hub."),
-            )
-    }
-
-    pub fn execute(&self, quiet: bool) -> Result<(), String> {
-        let data = json!({
-            "name": self.name,
-            "connection": self.connection,
-            "hub_name": self.hub_name,
-        });
-
-        create_function(self.name, "eventhub.rs", &data, quiet)
-    }
-}
-
-impl<'a> From<&'a ArgMatches<'a>> for EventHub<'a> {
-    fn from(args: &'a ArgMatches<'a>) -> Self {
-        EventHub {
-            name: args.value_of("name").unwrap(),
-            connection: args.value_of("connection").unwrap(),
-            hub_name: args.value_of("hub_name").unwrap_or(""),
         }
     }
 }
